Extract sidebar into separate class
