+++ src/sidebar/fzy.js
